NXP-25168: keep deprecated methods in deprecated interface
