+++ nuxeo-core/nuxeo-core-api/src/main/java/org/nuxeo/ecm/core/api/trash/TrashService.java
@@ -71,53 +71,6 @@ public interface TrashService {
     boolean isTrashed(CoreSession session, DocumentRef doc);
 
     /**
-     * Can a child of the folder be trashed?
-     *
-     * @param folder the folder
-     * @return {@code true} if the folder allows its children to be trashed
-     * @deprecated since 10.1 only used in JSF part, no replacement
-     */
-    @Deprecated
-    boolean folderAllowsDelete(DocumentModel folder);
-
-    /**
-     * Is at least one doc deletable according to its container?
-     *
-     * @param docs the documents
-     * @return {@code true} if one doc is in a folder that allows its children to be trashed
-     * @deprecated since 10.1 only used in JSF part, no replacement
-     */
-    @Deprecated
-    boolean checkDeletePermOnParents(List<DocumentModel> docs);
-
-    /**
-     * Is at least one doc deletable?
-     *
-     * @param docs the documents
-     * @param principal the current user (to check locks)
-     * @param checkProxies {@code true} to count proxies as non-deletable
-     * @return {@code true} if at least one doc is deletable
-     * @deprecated since 10.1 only used in JSF part, no replacement
-     */
-    @Deprecated
-    boolean canDelete(List<DocumentModel> docs, Principal principal, boolean checkProxies);
-
-    /**
-     * Are all documents purgeable/undeletable?
-     * <p>
-     * Documents need to be in the trash for this to be true, in addition to the standard permission checks.
-     *
-     * @param docs the documents
-     * @param principal the current user (to check locks)
-     * @return {@code true} if the documents are purgeable/undeletable
-     * @deprecated since 10.1, use {@link #canPurgeOrUntrash(List, Principal)}
-     */
-    @Deprecated
-    default boolean canPurgeOrUndelete(List<DocumentModel> docs, Principal principal) {
-        return canPurgeOrUntrash(docs, principal);
-    }
-
-    /**
      * Are all documents purgeable/undeletable?
      * <p>
      * Documents need to be in the trash for this to be true, in addition to the standard permission checks.
@@ -140,33 +93,6 @@ public interface TrashService {
     default boolean canPurgeOrUntrash(DocumentModel doc, Principal principal) {
         return canPurgeOrUntrash(Collections.singletonList(doc), principal);
     }
-
-    /**
-     * Gets the trash info for a list of documents.
-     *
-     * @param docs the documents
-     * @param principal the current user (to check locks)
-     * @param checkProxies {@code true} to count proxies as non-deletable
-     * @param checkDeleted {@code true} if documents have to be in the trashed state to be considered (otherwise
-     *            forbidden)
-     * @return the trash info
-     * @deprecated since 10.1 only used in JSF part, no replacement
-     */
-    @Deprecated
-    TrashInfo getTrashInfo(List<DocumentModel> docs, Principal principal, boolean checkProxies, boolean checkDeleted);
-
-    /**
-     * Gets the closest document's ancestor above all the paths.
-     * <p>
-     * This is used to find what safe document to redirect to when deleting some.
-     *
-     * @param doc the document
-     * @param paths the paths
-     * @return the closer document above doc and above all the paths
-     * @deprecated since 10.1 only used in JSF part, use {@link #getAboveDocument(DocumentModel, Principal)} instead.
-     */
-    @Deprecated
-    DocumentModel getAboveDocument(DocumentModel doc, Set<Path> paths);
 
     /**
      * Gets the first non trashed ancestor.
@@ -221,27 +147,13 @@ public interface TrashService {
     void purgeDocumentsUnder(DocumentModel parent);
 
     /**
-     * Undeletes documents (and ancestors if needed to make them visible).
-     * <p>
-     * Also fires async events to untrash the children.
-     *
-     * @param docs the documents to undelete
-     * @return the set of ancestors whose children have been untrashed (for UI notification)
-     * @deprecated since 10.1 use {@link #untrashDocuments(List)} instead
-     */
-    @Deprecated
-    Set<DocumentRef> undeleteDocuments(List<DocumentModel> docs);
-
-    /**
      * Unmoves documents from the trash.
      * <p>
      * Also fires async events to untrash the children.
      *
      * @param docs the documents to untrash
      */
-    default void untrashDocuments(List<DocumentModel> docs) {
-        undeleteDocuments(docs);
-    }
+    void untrashDocuments(List<DocumentModel> docs);
 
     /**
      * Unmoves document from the trash.
