+++ nuxeo-dm/nuxeo-platform-webapp-core/src/main/java/org/nuxeo/ecm/webapp/action/DeleteActionsBean.java
@@ -226,7 +226,7 @@ public class DeleteActionsBean implements DeleteActions, Serializable {
         if (docs == null) {
             return null;
         }
-        TrashInfo info = (TrashInfo) getTrashService().getTrashInfo(docs, currentUser, false, false);
+        TrashInfo info = getTrashService().getTrashInfo(docs, currentUser, false, false);
 
         DocumentModel targetContext = getTrashService().getAboveDocument(navigationContext.getCurrentDocument(),
                 info.rootPaths);
